Implement class World
