+++ projects/PathosEngine/PathosEngine.vcxproj.filters
@@ -282,6 +282,9 @@
     <ClCompile Include="src\badger\math\rotator.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
+    <ClCompile Include="src\pathos\actor\world.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\pathos\mesh\geometry.h">
@@ -621,6 +624,9 @@
       <Filter>헤더 파일</Filter>
     </ClInclude>
     <ClInclude Include="src\badger\types\os.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
+    <ClInclude Include="src\pathos\actor\world.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
   </ItemGroup>
