+++ projects/PathosEngine/src/pathos/actor/actor.h
@@ -9,6 +9,7 @@
 namespace pathos {
 
 	class Scene;
+	class World;
 	class ActorComponent;
 	class SceneComponent;
 
@@ -16,6 +17,7 @@ namespace pathos {
 	class Actor
 	{
 		friend class Scene;
+		friend class World;
 
 	protected:
 		Actor() { isInConstructor = true; } // Use Scene::spawnActor()
@@ -27,7 +29,7 @@ namespace pathos {
 		void registerComponent(ActorComponent* component);
 		void unregisterComponent(ActorComponent* component);
 
-		inline Scene* getOwnerScene() const { return owner; }
+		inline World* getWorld() const { return owner; }
 
 		inline SceneComponent* getRootComponent() { return rootComponent; }
 		vector3 getActorLocation() const;
@@ -65,7 +67,7 @@ namespace pathos {
 	private:
 		void fixRootComponent();
 
-		Scene* owner = nullptr;
+		World* owner = nullptr;
 		bool isInConstructor = false;
 		bool markedForDeath = false;
