+++ projects/PathosEngine/src/pathos/engine.cpp
@@ -1,11 +1,13 @@
 #include "engine.h"
 #include "engine_version.h"
 #include "console.h"
-#include "render/render_device.h"
-#include "render/render_deferred.h"
-#include "util/log.h"
-#include "util/resource_finder.h"
-#include "util/renderdoc_integration.h"
+#include "pathos/actor/world.h"
+#include "pathos/scene/scene.h"
+#include "pathos/render/render_device.h"
+#include "pathos/render/render_deferred.h"
+#include "pathos/util/log.h"
+#include "pathos/util/resource_finder.h"
+#include "pathos/util/renderdoc_integration.h"
 
 #include "pathos/loader/imageloader.h"    // subsystem: image loader
 #include "pathos/text/font_mgr.h"         // subsystem: font manager
@@ -53,8 +55,7 @@ namespace pathos {
 		: renderProxyAllocator(RENDER_PROXY_MEMORY)
 		, elapsed_gameThread(0.0f)
 		, elapsed_renderThread(0.0f)
-		, scene(nullptr)
-		, camera(nullptr)
+		, currentWorld(nullptr)
 		, render_device(nullptr)
 		, renderer(nullptr)
 		, timer_query(0)
@@ -313,10 +314,15 @@ namespace pathos {
 		return false;
 	}
 
-	void Engine::setWorld(Scene* inScene, Camera* inCamera)
+	void Engine::setWorld(World* inWorld)
 	{
-		scene = inScene;
-		camera = inCamera;
+		if (currentWorld != nullptr) {
+			currentWorld->destroy();
+			delete currentWorld;
+		}
+
+		currentWorld = inWorld;
+		currentWorld->initialize();
 	}
 
 	void Engine::tick()
@@ -328,8 +334,8 @@ namespace pathos {
 
 		// #todo-fps: This is wrong. Rendering rate should be also controlled...
 		if (maxFPS.getValue() > 0 && deltaSeconds < 1.0f / maxFPS.getValue()) {
-			if (scene != nullptr) {
-				scene->createRenderProxy();
+			if (currentWorld != nullptr) {
+				currentWorld->getScene().createRenderProxy();
 			}
 			return;
 		}
@@ -338,17 +344,9 @@ namespace pathos {
 
 		inputSystem->tick();
 
-		if (scene != nullptr) {
-			scene->tick(deltaSeconds);
-		}
-
-		auto callback = Engine::conf.tick;
-		if (callback != nullptr) {
-			callback(deltaSeconds);
-		}
-
-		if (scene != nullptr) {
-			scene->createRenderProxy();
+		if (currentWorld != nullptr) {
+			currentWorld->tick(deltaSeconds);
+			currentWorld->getScene().createRenderProxy();
 		}
 
 		elapsed_gameThread = stopwatch_gameThread.stop() * 1000.0f;
@@ -377,8 +375,8 @@ namespace pathos {
 		// #todo-cmd-list: deferred command lists here
 
 		// Renderer adds more immediate commands
-		if (renderer && scene && camera) {
-			renderer->render(immediateContext, scene, camera);
+		if (renderer && currentWorld) {
+			renderer->render(immediateContext, &currentWorld->getScene(), &currentWorld->getCamera());
 			immediateContext.flushAllCommands();
 		}
 
@@ -400,8 +398,8 @@ namespace pathos {
 
 		renderProxyAllocator.clear();
 
-		if (scene != nullptr) {
-			scene->clearRenderProxy();
+		if (currentWorld != nullptr) {
+			currentWorld->getScene().clearRenderProxy();
 		}
 
 		elapsed_renderThread = stopwatch_renderThread.stop() * 1000.0f;
