+++ projects/PathosEngine/src/pathos/scene/scene.cpp
@@ -1,54 +1,11 @@
 #include "scene.h"
 #include "pathos/console.h"
+#include "pathos/actor/world.h"
 #include "pathos/actor/scene_component.h"
 #include "pathos/light/point_light_component.h"
 #include "pathos/light/directional_light_component.h"
 
 namespace pathos {
-
-	static ConsoleVariable<int32_t> cvar_visualize_depth("r.visualize_depth", 0, "visualize the scene depth");
-
-	void Scene::destroyActor(Actor* actor) {
-		CHECKF(actor != nullptr, "Parameter is null");
-		CHECKF(actor->owner == this, "this Actor does not belong to this Scene");
-
-		int32 ix = -1;
-		for (int32 i = 0; i < actors.size(); ++i) {
-			if (actors[i] == actor) {
-				ix = i;
-				break;
-			}
-		}
-
-		if (ix == -1) {
-			// Should fix if enters here
-			CHECK_NO_ENTRY();
-		} else {
-			actor->markedForDeath = true;
-			actorsToDestroy.push_back(actors[ix]);
-			actors.erase(actors.begin() + ix);
-		}
-	}
-
-	void Scene::tick(float deltaSeconds) {
-		// Destroy actors that were marked for death
-		{
-			for (auto& actor : actorsToDestroy) {
-				actor->onDestroy();
-				delete actor;
-			}
-			actorsToDestroy.clear();
-		}
-
-		// Tick
-		{
-			for (auto& actor : actors) {
-				if (!actor->markedForDeath) {
-					actor->onTick(deltaSeconds);
-				}
-			}
-		}
-	}
 
 	void Scene::clearRenderProxy() {
 		proxyList_directionalLight.clear();
@@ -61,7 +18,7 @@ namespace pathos {
 	}
 
 	void Scene::createRenderProxy() {
-		for (auto& actor : actors) {
+		for (auto& actor : getWorld()->actors) {
 			if (!actor->markedForDeath) {
 				for (ActorComponent* actorComponent : actor->components) {
 					actorComponent->createRenderProxy(this);
