+++ projects/PathosEngine/src/pathos/scene/scene_capture_component.cpp
@@ -1,5 +1,7 @@
 #include "scene_capture_component.h"
 #include "scene.h"
+#include "pathos/actor/world.h"
+#include "pathos/actor/actor.h"
 #include "pathos/render/render_deferred.h"
 #include "pathos/render/render_device.h"
 #include "pathos/render/render_target.h"
@@ -23,28 +25,27 @@ namespace pathos {
 			renderer->setFinalRenderTarget(renderTarget);
 		}
 
-		Scene* scene = getOwner()->getOwnerScene();
+		Scene& scene = getOwner()->getWorld()->getScene();
 
 		const float aspectRatio = (float)renderTarget->getWidth() / (float)renderTarget->getHeight();
 
 		// #todo-renderer: Let Renderer::render() take a view family parameter, not a camera.
-		PerspectiveLens lens(fovY, aspectRatio, zNear, zFar);
-		Camera tempCamera(&lens);
+		Camera tempCamera(PerspectiveLens(fovY, aspectRatio, zNear, zFar));
 		tempCamera.moveToPosition(getLocation());
 
 		Rotator componentRotation = getRotation();
 		tempCamera.setYaw(componentRotation.yaw);
 		tempCamera.setPitch(componentRotation.pitch);
 
-		scene->createRenderProxy();
+		scene.createRenderProxy();
 
 		ScopedGpuCounter::enable = false;
 
 		ENQUEUE_RENDER_COMMAND(
-			[renderer, scene, &tempCamera](RenderCommandList& cmdList) {
+			[renderer, &scene, &tempCamera](RenderCommandList& cmdList) {
 				SCOPED_DRAW_EVENT(SceneCapture);
 
-				renderer->render(cmdList, scene, &tempCamera);
+				renderer->render(cmdList, &scene, &tempCamera);
 			}
 		);
 
@@ -52,7 +53,7 @@ namespace pathos {
 		FLUSH_RENDER_COMMAND();
 
 		ENQUEUE_RENDER_COMMAND(
-			[renderer, scene, &tempCamera](RenderCommandList& cmdList) {
+			[renderer](RenderCommandList& cmdList) {
 				// This destroys GL objects immediately, before render commands generated by renderer->render() are executed.
 				renderer->releaseResources(cmdList);
 			}
@@ -63,7 +64,7 @@ namespace pathos {
 
 		ScopedGpuCounter::enable = true;
 
-		scene->clearRenderProxy();
+		scene.clearRenderProxy();
 		
 		delete renderer;
