+++ projects/Test_DeferredRenderer/Test_DeferredRenderer.vcxproj.filters
@@ -21,9 +21,15 @@
     <ClCompile Include="src\csm_debugger.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
+    <ClCompile Include="src\world1.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\csm_debugger.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
+    <ClInclude Include="src\world1.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
   </ItemGroup>
