+++ projects/Test_DeferredRenderer/src/csm_debugger.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "badger/types/vector_types.h"
+#include "pathos/camera/camera.h"
 #include "pathos/mesh/static_mesh_actor.h"
 #include "pathos/mesh/geometry_procedural.h"
 using namespace pathos;
