+++ projects/Test_SkeletalAnimation/Test_SkeletalAnimation.vcxproj.filters
@@ -24,6 +24,9 @@
     <ClCompile Include="src\skinned_mesh.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
+    <ClCompile Include="src\world2.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\daeloader.h">
@@ -36,6 +39,9 @@
       <Filter>헤더 파일</Filter>
     </ClInclude>
     <ClInclude Include="src\animation.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
+    <ClInclude Include="src\world2.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
   </ItemGroup>
